Add a bootstrap build configuration.
This configuration can be used to build a bootstrap variant of U-Boot,
which is a full build but is linked to load at address 0x00000000 and
does low-level hardware initialization. Furthermore, this variant never
boots automatically. It can be used to bootstrap a board via the UART5
service boot mode.