Sound: Tegra: Add sound driver for T114/T124 audio
AHUB is the audio hub on Tegra, and controls the audio/apbif
settings for simple audio playback ('sound init/play'). I2S is
obvious, and sets up the I2S controller in Tegra for 16-bit
stereo PCM data transfer from the AHUB FIFO to the audio codec.
Tegra-sound is the overall driver that sets up the codec regs,
creates the 400Hz square wave data, and then plays it back.

This patch is based on a patch from Tom Wareen for chromium:
Reviewed-on: https://chromium-review.googlesource.com/175343

Signed-off-by: Julian Scheel <julian@jusst.de>