ARM: tegra20: Add a missing entry in the pullid enum
It seems two entries were merged in one when this file has been
created. The GPSLXAU entries is obviously a mix of GPU and SLXA which
are next to each other according to the datasheet. Moreover it can be
noticed because the APB_MISC_PP_PULLUPDOWN_REG_B_0 register only have
15 entries instead of 16.

Also fix the pin group descriptions that were using these buggy
entries. In particular SLXA that needed to used CRTP to actually
write the SLXA register.

Change-Id: I7080258a0f9faba656b02632b372792ba0f30642
Signed-off-by: Alban Bedel <alban.bedel@avionic-design.de>
Acked-by: Stephen Warren <swarren@nvidia.com>
Reviewed-on: http://review.adnet.avionic-design.de/989
Reviewed-by: Soeren Grunewald <soeren.grunewald@avionic-design.de>