meerkat: Remove global GPIO initialisation
The bootloader makes a global pin initialisation, configuring nearly all
as gpios. Because the bootloader is a common bootloader and
independent from the baseboard as well, this initialisation does not
really make sense.
In addition this can cause problems on the kernel side, because if the
pin is set to gipo, it is not trivial to switch it to another function.
Therefore this removes the initialisation call and the gpio pin array.

Change-Id: Ifb9a1ebca6879247d3c19d049c38193279a1c54f
Signed-off-by: Nicolas Bänisch <nicolas.baenisch@avionic-design.de>
Reviewed-on: http://review.adnet.avionic-design.de/3226
Reviewed-by: Alban Bedel <alban.bedel@avionic-design.de>