kein-baseboard: Add a command to update u-boot
Updating the u-boot image require a few steps: selecting the proper
mmc device, converting the size from bytes to mmc blocks and finally
writing it to the device. This add a command that does all at once,
the user just need to load the image using a standard load command
(loadb, tftp, dhcp, etc) and run update_uboot.

This also enable the setexpr command as it is needed to compute the
number of blocks from the image size.

Change-Id: Ib30f40506974c1d19da95e45d13df8830d9453f1
Signed-off-by: Alban Bedel <alban.bedel@avionic-design.de>
Reviewed-on: http://review.adnet.avionic-design.de/6236
Reviewed-by: Nikolaus Schulz <nikolaus.schulz@avionic-design.de>